requirements.txt
pandas
numpy
seaborn
scikit-learn
kaggle
kagglehub
astrapy
python-dotenv
time
flask


-e .